changed the path with respect to new components
