+++ src/components/pages/learn/comparisons/Prefooter.js
