+++ src/components/pages/learn/overview/Hero/index.js
@@ -4,7 +4,7 @@ import { Flex, Box } from 'serverless-design-system/src'
 import { HeroWrapper } from 'src/fragments'
 import HeroLeftSection from './LeftSection'
 import HeroRightSection from './RightSection'
-import LearnOverviewBg from '../../../../assets/images/learnOverviewBg.png'
+import LearnOverviewBg from '../../../../../assets/images/learnOverviewBg.png'
 
 const DashboardHero = () => (
   <HeroWrapper
