+++ src/pages/learn/use-cases.js
@@ -2,9 +2,9 @@ import React from 'react'
 
 import DefaultLayout from 'src/layouts/Default'
 
-import Prefooter from './Prefooter'
-import Hero from './Hero'
-import UseCaseScrollList from './UseCaseScrollList'
+import Prefooter from 'src/components/pages/learn/use-cases/Prefooter'
+import Hero from 'src/components/pages/learn/use-cases/Hero'
+import UseCaseScrollList from 'src/components/pages/learn/use-cases/UseCaseScrollList'
 
 const UseCases = () => (
   <DefaultLayout prefooter={Prefooter} transparentHeader>
